requirements.txt
accelerate==1.11.0
acres==0.5.0
adjustText==1.3.0
ai-data-science-team==0.0.0.9016
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
antropy==0.1.9
anyio==4.11.0
arch==7.2.0
arxiv==2.2.0
asgiref==3.9.2
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
build==1.3.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
distro==1.9.0
docling==2.57.0
docling-core==2.49.0
docling-ibm-models==3.10.0
docling-parse==4.7.0
durationpy==0.10
easyocr==1.7.2
emoji==2.15.0
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.1
faiss-cpu==1.12.0
Faker==37.11.0
fastapi==0.115.12
fastjsonschema==2.21.2
fastmcp==0.4.1
feedparser==6.0.12
filelock==3.19.1
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
holidays==0.81
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
json_repair==0.39.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.3.27
langchain-anthropic==0.3.10
langchain-chroma==0.2.3
langchain-community==0.3.31
langchain-core==0.3.79
langchain-experimental==0.3.4
langchain-mcp-adapters==0.0.6
langchain-openai==0.3.11
langchain-tavily==0.1.5
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.3.20
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.74
langgraph-supervisor==0.0.22
langgraph-swarm==0.0.11
langsmith==0.4.37
latex2mathml==3.78.1
lazy_loader==0.4
llvmlite==0.45.0
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mcp==1.4.1
mdurl==0.1.2
missingno==0.5.2
mizani==0.13.5
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.6.0
nbformat==5.10.4
neo4j==5.28.1
neo4j-graphrag==1.8.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
nltk==3.9.2
numba==0.62.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnxruntime==1.23.0
openai==1.69.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandas_flavor==0.7.0
parso==0.8.5
pathlib==1.0.1
pathos==0.3.4
pathspec==0.12.1
patsy==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pi_heif==1.1.1
pillow==11.3.0
platformdirs==4.4.0
plotly==6.1.1
plotnine==0.14.5
pluggy==1.6.0
polars==1.33.1
polyfactory==2.22.3
posthog==6.7.6
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
prov==2.1.1
psutil==7.1.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.5
pyparsing==3.2.5
pypdf==5.9.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytimetk==1.1.2
pytz==2025.2
pywin32==311
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
rank-bm25==0.2.2
RapidFuzz==3.14.1
rapidocr==3.4.2
rdflib==7.2.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semchunk==2.2.2
sentence-transformers==2.7.0
serpapi==0.1.5
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.40
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.5
streamlit==1.50.0
supersmoother==0.4
sweetviz==2.1.4
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.5.3
tenacity==9.0.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.9.0
timebasedcv==0.3.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.57.1
tsfeatures==0.4.5
typer==0.19.2
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.3
xarray==2025.6.1
xgboost==3.0.5
xlsxwriter==3.2.9
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
